An additional base64_encode( hash_hmac( prefix, true ) ) was added
This would shorten the $local_part of the email id. We have constraints
that the $local_part be < 64 chars. This would decrease the size of the
hash from 32 to 24.
* Also fixed minor issues with the BounceHandlerActions. The bouncePeriod
was not properly checked for earlier. Now the bouncehandler considers only
the required bounces in the bounceValidPerioid
* Made the insertion to database depend on the Timestamp extracted from the
VERP address rather than the incsecure email headers, which might be tampered
in between.
* Removed unwanted variables.
* Improved the regex functions

Bug: 69099
Change-Id: I0f9b098982adf023347e3f1343f1e16ccec91df1