Removed Wikimedia specific configuratoins from BounceHandler extension
As per the plan, the bouncehandler extension will get installed in all
WMF wikis, and having $wgDomainPart = 'meta.wikimedia.org' will make it
difficult later as each of them needs to get changed to $wgServerName.
It is important that VERP address from a wiki have the wiki domain as $domain
part. Now, the $domain part defaults to $wgServerName, which can be changed later.

Bug: 73043
Change-Id: I8bdbac798e62be527eb90781dd89636cc9290fe1