Improve default $wgUnrecognizedBounceNotify behavior
If this variable is not set, don't set it to the no-reply@ address, because
that address most likely isn't monitored and will be blackholed. Just
skip sending an email in that case - if the sysadmin wants notifications,
they should explicitly configure them.

And clean up $wgUnrecognizedBounceNotify to be an array by default, not null.

Bug: T180192
Change-Id: I740e702770d8add447aa419d32d48977fd269b27